feat: plugin system (v0.3.0)
- isolated source and destination connectors to dedicated subpackages
