+++ src/hermes/exceptions.py
@@ -41,32 +41,6 @@ class DestinationError(HermesError):
         return f"{self.connector_name}: {self.error_message}"
 
 
-class AthenaIcebergDestinationError(DestinationError):
-    def __init__(self, glue_database, glue_table, table_location, temp_path, error):
-        self.connector_name = "Athena Iceberg"
-        self.error_message = f"""error while loading data to {glue_database}.{glue_table}.
-            table location: {table_location}, temp path: {temp_path}
-            error : {error}
-        """
-        super().__init__(self.connector_name, self.error_message)
-
-
-class ObjectStorageDestinationError(DestinationError):
-    def __init__(self, name, path, error):
-        self.connector_name = "Object Storage"
-        self.error_message = f"""error while loading data to {name}, located at {path}.
-            error : {error}
-        """
-        super().__init__(self.connector_name, self.error_message)
-
-
-class CustomSourceError(SourceError):
-    def __init__(self, name, process_step, error):
-        self.connector_name = "Custom Source"
-        self.error_message = f"""error during {process_step} for source {name}.
-            error : {error}
-        """
-        super().__init__(self.connector_name, self.error_message)
 
 
 class PipelineError(HermesError):
