+++ src/plugins/destinations/athena_iceberg/pyproject.toml
@@ -1,0 +1,24 @@
+[project]
+name = "athena_iceberg_destination"
+version = "0.1.0"
+description = "Hermes Athena Iceberg destination connector"
+readme = "README.md"
+authors = [
+    { name = "Brocoli Data", email = "contact.brocoli@gmail.com" }
+]
+requires-python = ">=3.12"
+dependencies = [
+    "hermes",
+]
+
+[tool.uv.sources]
+# hermes = { path = "../../.." }
+# generic_object_storage_destination = { path = "../generic_object_storage"}
+
+[dependency-groups]
+dev = [
+    "awswrangler>=3.10.1",
+]
+
+[project.entry-points."hermes_plugins.destinations"]
+athena_iceberg = "athena_iceberg_destination.connector:AthenaIcebergDestination"
