+++ src/plugins/destinations/generic_object_storage/pyproject.toml
@@ -1,0 +1,16 @@
+[project]
+name = "generic_object_storage_destination"
+version = "0.1.0"
+description = "Hermes generic object storage destination connector"
+readme = "README.md"
+authors = [
+    { name = "Brocoli Data", email = "contact.brocoli@gmail.com" }
+]
+requires-python = ">=3.12"
+dependencies = [
+    "hermes",
+    "fsspec>=2025.2.0",
+]
+
+[tool.uv.sources]
+# hermes = { path = "../../.." }
