+++ src/plugins/destinations/s3/pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "s3_destination"
+version = "0.1.0"
+description = "Hermes s3 destination connector"
+readme = "README.md"
+authors = [
+    { name = "Brocoli Data", email = "contact.brocoli@gmail.com" }
+]
+requires-python = ">=3.12"
+dependencies = [
+    "hermes",
+    "generic_object_storage_destination",
+    "s3fs>=2025.2.0",
+]
+
+[tool.uv.sources]
+# hermes = { path = "../../.." }
+generic_object_storage_destination = { path = "../generic_object_storage"}
+
+[project.entry-points."hermes_plugins.destinations"]
+s3 = "s3_destination.connector:S3Destination"
