+++ src/tests/unit_tests/destination/test_athena_iceberg.py
@@ -2,11 +2,10 @@ from unittest.mock import patch
 
 import pandas as pd
 import pytest
+from athena_iceberg_destination.connector import AthenaIcebergDestination
+from athena_iceberg_destination.exceptions import AthenaIcebergDestinationError
 from botocore.exceptions import BotoCoreError
 from omegaconf import DictConfig
-
-from hermes.destinations.athena_iceberg import AthenaIcebergDestination
-from hermes.exceptions import AthenaIcebergDestinationError
 
 
 @pytest.fixture
@@ -30,6 +29,7 @@ def test_dataframe():
     return pd.DataFrame({"id": [1, 2, 3], "value": ["A", "B", "C"]})
 
 
+@pytest.mark.athena_iceberg_destination
 def test_load_data(mock_athena_to_iceberg, test_destination_config, test_dataframe):
     """Test the load method of AthenaIcebergDestination."""
     destination = AthenaIcebergDestination(test_destination_config)
@@ -49,6 +49,7 @@ def test_load_data(mock_athena_to_iceberg, test_destination_config, test_datafra
     )
 
 
+@pytest.mark.athena_iceberg_destination
 @patch("awswrangler.athena.to_iceberg", side_effect=BotoCoreError)
 def test_load_data_exception(
     mock_athena_to_iceberg, test_destination_config, test_dataframe
